change all size_t to uint16_t

+++ sensing/pointcloud_preprocessor/src/filter.cpp
@@ -70,7 +70,7 @@ pointcloud_preprocessor::Filter::Filter(
   {
     tf_input_frame_ = declare_parameter<std::string>("input_frame");
     tf_output_frame_ = declare_parameter<std::string>("output_frame");
-    max_queue_size_ = declare_parameter<int>("max_queue_size");
+    max_queue_size_ = declare_parameter<uint16_t>("max_queue_size");
 
     // ---[ Optional parameters
     use_indices_ = declare_parameter<bool>("use_indices");
